\encoding{UTF-8}

\name{angle_count_cpp}

\alias{angle_count_cpp}

\title{Calculate dominant diameters and heights for simulations for angle-count plots.}

\description{Calculate dominant diameters and heights for simulations for angle-count plots.}

\usage{
angle_count_cpp(baf_seq, baf, d, h, num)}

\arguments{

  \item{baf_seq}{BAF sequence for simulations}

  \item{baf}{BAF masimum for simulations}

  \item{d}{vector with diameters opf trees}

  \item{h}{vector with heights of trees}

  \item{num}{number of trees considered for calculating diameter and dominant height}}

\value{Data frame.}


\author{
Juan Alberto Molina-Valero,
María José Ginzo Villamayor,
Manuel Antonio Novo Pérez,
Adela Martínez-Calvo,
Juan Gabriel Álvarez-González,
Fernando Montes and
César Pérez-Cruzado.}

\keyword{internal}



