\encoding{UTF-8}

\name{chunk_size_from_area}

\alias{chunk_size_from_area}

\title{This function compute the chunk size from the area of the las-catalog}

\description{This function compute the chunk size from the area of the las-catalog}

\usage{
chunk_size_from_area(las_catalog, num_parts)}

\arguments{

  \item{las_catalog}{las_catalog}

  \item{num_parts}{num_parts}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



