\encoding{UTF-8}

\name{datatable_grid}

\alias{datatable_grid}

\title{This function create overlapping polygons using the data.table R package}

\description{This function create overlapping polygons using the data.table R package}

\usage{
datatable_grid(data_inp, dist = dist, n_grid = 20, verify_visually = FALSE)}

\arguments{

  \item{data_inp}{data_inp}

  \item{dist}{dist}

  \item{n_grid}{n_grid}

  \item{verify_visually}{verify_visually}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



