\encoding{UTF-8}

\name{fit_circle_cpp_modified}

\alias{fit_circle_cpp_modified}

\title{This function fit a circle based on 3 points}

\description{This function fit a circle based on 3 points}


\usage{
fit_circle_cpp_modified(points)}

\arguments{

  \item{points}{points}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



