\encoding{UTF-8}

\name{geometric_features_py}

\alias{geometric_features_py}

\title{This function obtains geometric features at point level}

\description{This function obtains geometric features at point level}


\usage{
geometric_features_py(data,
                      dist = 0.1,
                      geom_features_input = NULL,
                      threads = 1)}

\arguments{

  \item{data}{data}

  \item{dist}{dist}

  \item{geom_features_input}{geom_features_input}

  \item{threads}{threads}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



