\encoding{UTF-8}

\name{height_perc_cpp}

\alias{height_perc_cpp}

\title{Calculate dominant diameters and heights for simulations for angle-count plots.}

\description{Calculate dominant diameters and heights for simulations for angle-count plots.}

\usage{
  height_perc_cpp(rho_seq, z, rho)}

\arguments{

  \item{rho_seq}{BAF sequence for simulations}

  \item{z}{vector with heights of trees}

  \item{rho}{number of trees considered for calculating diameter and dominant height}}

\value{Data frame.}


\author{
  Juan Alberto Molina-Valero,
  María José Ginzo Villamayor,
  Manuel Antonio Novo Pérez,
  Adela Martínez-Calvo,
  Juan Gabriel Álvarez-González,
  Fernando Montes and
  César Pérez-Cruzado.}

\keyword{internal}
