% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fortls_python_deps.R
\name{install_fortls_python_deps}
\alias{install_fortls_python_deps}
\title{Install Python dependencies required by FORTLS}
\usage{
install_fortls_python_deps(envname = "fortls_env", method = "auto")
}
\arguments{
\item{envname}{Name of the virtual environment or conda environment.}

\item{method}{Either "virtualenv", "conda", or "auto".}
}
\description{
Install Python dependencies required by FORTLS
}
