\encoding{UTF-8}

\name{internal_ransac}

\alias{internal_ransac}

\title{Apply RANSAC algorithm to estimate diameters.}

\description{Apply RANSAC algorithm to estimate diameters.}

\usage{
internal_ransac(dat,
                dist = 0.05)}

\arguments{

  \item{dat}{dat}

  \item{dist}{dist}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



