\encoding{UTF-8}

\name{is_one_row_all_na}

\alias{is_one_row_all_na}

\title{This function was updated to return also the input data with the computed voxels}

\description{This function was updated to return also the input data with the computed voxels}


\usage{
is_one_row_all_na(df)}

\arguments{

  \item{df}{df}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



