\encoding{UTF-8}

\name{iterations_RANSAC}

\alias{iterations_RANSAC}

\title{Function that performs the "RANSAC_cpp" N-times}

\description{Function that performs the "RANSAC_cpp" N-times}


\usage{
iterations_RANSAC(data, n_iterations = 600)}

\arguments{

  \item{data}{data}

  \item{n_iterations}{n_iterations}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



