\encoding{UTF-8}

\name{k_tree_cpp}

\alias{k_tree_cpp}

\title{Calculate dominant diameters and heights for simulations for angle-count plots.}

\description{Calculate dominant diameters and heights for simulations for angle-count plots.}

\usage{
  k_tree_cpp(k_seq, radius_seq, k, d, h, num)}

\arguments{

  \item{k_seq}{BAF sequence for simulations}

  \item{radius_seq}{BAF sequence for simulations}

  \item{k}{BAF masimum for simulations}

  \item{d}{vector with diameters opf trees}

  \item{h}{vector with heights of trees}

  \item{num}{number of trees considered for calculating diameter and dominant height}}

\value{Data frame.}


\author{
  Juan Alberto Molina-Valero,
  María José Ginzo Villamayor,
  Manuel Antonio Novo Pérez,
  Adela Martínez-Calvo,
  Juan Gabriel Álvarez-González,
  Fernando Montes and
  César Pérez-Cruzado.}

\keyword{internal}

