\encoding{UTF-8}

\name{ncr_point_cloud_double}

\alias{ncr_point_cloud_double}

\title{Calculate dominant diameters and heights for simulations for angle-count plots.}

\description{Calculate dominant diameters and heights for simulations for angle-count plots.}

\usage{
  ncr_point_cloud_double(m)}

\arguments{

  \item{m}{matrix with points.}}

\value{Data frame.}


\author{
  Juan Alberto Molina-Valero,
  María José Ginzo Villamayor,
  Manuel Antonio Novo Pérez,
  Adela Martínez-Calvo,
  Juan Gabriel Álvarez-González,
  Fernando Montes and
  César Pérez-Cruzado.}

\keyword{internal}
