\encoding{UTF-8}

\name{python_folder}

\alias{python_folder}

\title{Define the path to the folder containing the Python scripts relative to the package directory}

\description{Define the path to the folder containing the Python scripts relative to the package directory}


\usage{
python_folder("python", package = "FORTLS")}

\arguments{

  \item{package}{package}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



