\encoding{UTF-8}

\name{random.forest.fit}

\alias{random.forest.fit}

\title{Define the path to the folder containing the Python scripts relative to the package directory}

\description{Define the path to the folder containing the Python scripts relative to the package directory}


\usage{
random.forest.fit(dir.data, file_name = NULL)}

\arguments{

  \item{dir.data}{dir.data}

  \item{file_name}{file_name}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



