\encoding{UTF-8}

\name{random.forest.sp}

\alias{random.forest.sp}

\title{Define the path to the folder containing the Python scripts relative to the package directory}

\description{Define the path to the folder containing the Python scripts relative to the package directory}


\usage{
random.forest.sp(file_name)}

\arguments{

  \item{file_name}{file_name}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



