\encoding{UTF-8}

\name{RANSAC_cpp}

\alias{RANSAC_cpp}

\title{Apply RANSAC algorithm to estimate diameters.}

\description{Apply RANSAC algorithm to estimate diameters.}

\usage{
RANSAC_cpp(data, dist)}

\arguments{

  \item{data}{Points}

  \item{dist}{Distance}}

\value{Data frame.}


\author{
Juan Alberto Molina-Valero,
María José Ginzo Villamayor,
Manuel Antonio Novo Pérez,
Adela Martínez-Calvo,
Juan Gabriel Álvarez-González,
Fernando Montes and
César Pérez-Cruzado.}

\keyword{internal}



