\encoding{UTF-8}

\name{sample_indices}

\alias{sample_indices}

\title{Sample_indices}

\description{Sample_indices}


\usage{
sample_indices(total_number_obs, return_obs)}

\arguments{

  \item{total_number_obs}{total_number_obs}

  \item{return_obs}{return_obs}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



