\encoding{UTF-8}

\name{save_file_as_laz}

\alias{save_file_as_laz}

\title{Save a file as a .laz file}

\description{Save a file as a .laz file}


\usage{
save_file_as_laz(input_file, output_laz_file, threads, EPSG = 32633)}

\arguments{

  \item{input_file}{input_file}

  \item{output_laz_file}{output_laz_file}

  \item{threads}{threads}

  \item{EPSG}{EPSG}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



