\encoding{UTF-8}

\name{save_to_tiles}

\alias{save_to_tiles}

\title{Save to tiles and receive the metadata}

\description{Save to tiles and receive the metadata}


\usage{
save_to_tiles(las_catalog,
              chunk_size_input,
              dir_save_tiles,
              buffer_size = 2,
              increment_chunk_size_factor = 1.4)}

\arguments{

  \item{las_catalog}{las_catalog}

  \item{chunk_size_input}{chunk_size_input}

  \item{dir_save_tiles}{dir_save_tiles}

  \item{buffer_size}{buffer_size}

  \item{increment_chunk_size_factor}{increment_chunk_size_factor}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



