\encoding{UTF-8}

\name{sort_sublists}

\alias{sort_sublists}

\title{Secondary function to use and sort the sublists before using as input to the Rcpp function}

\description{Secondary function to use and sort the sublists before using as input to the Rcpp function}


\usage{
sort_sublists(lst_unsorted, MIN_PNTS_VX = 2)}

\arguments{

  \item{lst_unsorted}{lst_unsorted}

  \item{MIN_PNTS_VX}{MIN_PNTS_VX}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



