\encoding{UTF-8}

\name{species.classification}

\alias{species.classification}

\title{Species classification}

\description{Species classification}


\usage{
species.classification(data,
                       tree.tls,
                       dist = 0.1,
                       type = 3,
                       save.result = TRUE,
                       dir.result = NULL)}

\arguments{

  \item{data}{data}

  \item{tree.tls}{tree.tls}

  \item{dist}{dist}

  \item{type}{type}

  \item{save.result}{save.result}

  \item{dir.result}{dir.result}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



