\encoding{UTF-8}

\name{voxel_grid_downsampling}

\alias{voxel_grid_downsampling}

\title{Voxel down sampling}

\description{Voxel down sampling}


\usage{
voxel_grid_downsampling(points, voxel_size)}

\arguments{

  \item{points}{points}

  \item{voxel_size}{voxel_size}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



