\encoding{UTF-8}

\name{VoxR_vox_update}

\alias{VoxR_vox_update}

\title{This function was updated to return also the input data with the computed voxels}

\description{This function was updated to return also the input data with the computed voxels}

\usage{
RANSAC_cpp(data, res, full.grid, message)}

\arguments{

  \item{data}{data}

  \item{res}{res}

  \item{full.grid}{full.grid}

  \item{message}{message}}

\value{Data frame}


\author{
Juan Alberto Molina-Valero}

\keyword{internal}



