\encoding{UTF-8}

\name{weighted_mean_sqrt}

\alias{weighted_mean_sqrt}

\title{Calculate weighted quadratic mean.}

\description{Calculate weighted quadratic mean.}

\usage{

  weighted_mean_sqrt(x, w)}

\arguments{

  \item{x}{attribute for which calculate mean.}

  \item{w}{vector with weights associated to x.}}

\value{Data frame.}


\author{
  Juan Alberto Molina-Valero,
  María José Ginzo Villamayor,
  Manuel Antonio Novo Pérez,
  Adela Martínez-Calvo,
  Juan Gabriel Álvarez-González,
  Fernando Montes and
  César Pérez-Cruzado.}

\keyword{internal}
