\encoding{UTF-8}

\name{fit_circles_macif}

\alias{fit_circles_macif}

\title{Tree stem diameter estimation.}

\description{Tree stem diameter estimation.}

\usage{
  fit_circles_macif(points, step = 0.035, width = 0.07, ...)}

\arguments{

  \item{points}{points}

  \item{step}{step}

  \item{width}{width}

  \item{...}{...}}

\value{Data frame}

\references{

García-Pascual, B., Martín-Cortés, C., Zhou, X., Lopatin, E., Acuna, M., & Kärhä, K. (2025). Tree stem diameter estimation using inexpensive UAV photogrammetric data and Monte Carlo methods. \emph{ISPRS Annals of the Photogrammetry, Remote Sensing and Spatial Information Sciences}, \bold{10}, 49-56. \doi{https://doi.org/10.5194/isprs-annals-X-2-W2-2025-49-2025}.}

\author{
  Borja García-Pascual,
  and Juan Alberto Molina-Valero.}
